Many changes
Tons of changes in this. Now there's a dynamic heatmap by data. The data points changed - negative affect isn't there anymore, but can add it back in. The map is essentially functionless right now but I aim to make it more functional in the future. There is a heatmap and scatterplot that responds when you click on the heatmap. These are animated by year.

TODO: Get more data so that this framework actually makes sense. @Jonathan if you want to do this I'd love it and I can integrate in.